Added update_dyn_nuc_field
Added function to update the force on the nucleus from each pixel. This
function is completely different from the previous program to address
issues with protrusions pulling the nucleus out of the cell. The updated
version uses BFS to make sure force is propagated from within the cell,
and normalization is achieved via a sigmoid function restricting values
between 1 and 0 but ensuring values near the average have a propagation
probability of 0.5 (and retraction probability of 0.5) to make sure
values are balanced and prevents all values from being high or low