Fixed parallelization issues with protrude_nuc and added retract_nuc
There was an issue with the protrude_nuc function with writing to nuc_,
which is a sparse matrix, in a parallel section, which can cause the
function to break. Now all of the values to update are saved into a
vector before being added in the critical section to avoid issues.

retract_nuc() was also added by effectively doing the same thing as the
protrude_nuc function but some of the values are inverted.